requirements.txt
numpy==2.1.2
pynput==1.8.1
matplotlib==3.10.3
opencv-python==4.11.0.86
pillow==11.0.0
torch==2.7.0+cu128
torchvision==0.22.0+cu128
typing_extensions==4.12.2
filelock==3.13.1
fsspec==2024.6.1
jinja2==3.1.4
networkx==3.3 